Adding better tests
